docs/experiments/260424-auto-edit-experiment/requirements.txt
strands-agents>=1.1.0
strands-agents-tools>=0.2.2
bedrock-agentcore>=0.1.1
bedrock-agentcore-starter-toolkit>=0.1.2
boto3>=1.39.0
PyGithub>=2.1.0
PyYAML>=6.0
pytest>=8.0.0
python-dateutil>=2.9.0
